+++ cli/src/commands/git/mod.rs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 pub mod clone;
+pub mod colocate;
 pub mod export;
 pub mod fetch;
 pub mod import;
@@ -25,6 +26,8 @@ use clap::Subcommand;
 
 use self::clone::cmd_git_clone;
 use self::clone::GitCloneArgs;
+use self::colocate::cmd_git_colocate;
+use self::colocate::GitColocateArgs;
 use self::export::cmd_git_export;
 use self::export::GitExportArgs;
 use self::fetch::cmd_git_fetch;
@@ -52,6 +55,7 @@ use crate::ui::Ui;
 #[derive(Subcommand, Clone, Debug)]
 pub enum GitCommand {
     Clone(GitCloneArgs),
+    Colocate(GitColocateArgs),
     Export(GitExportArgs),
     Fetch(GitFetchArgs),
     Import(GitImportArgs),
@@ -70,6 +74,7 @@ pub fn cmd_git(
 ) -> Result<(), CommandError> {
     match subcommand {
         GitCommand::Clone(args) => cmd_git_clone(ui, command, args),
+        GitCommand::Colocate(args) => cmd_git_colocate(ui, command, args),
         GitCommand::Export(args) => cmd_git_export(ui, command, args),
         GitCommand::Fetch(args) => cmd_git_fetch(ui, command, args),
         GitCommand::Import(args) => cmd_git_import(ui, command, args),